requirements.txt
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
configparser==3.7.4
Flask
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-MySQLdb
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pip==19.1.1
pycparser==2.19
PyMySQL
python-crontab==2.3.6
python-dateutil==2.8.0
pywinusb==0.4.2
setuptools==41.0.1
six==1.12.0
SQLAlchemy==1.2.16
virtualenv==15.1.0
Werkzeug==0.14.1
WTForms==2.2.1
